Move new tabs to a separate component
