+++ libs/conversation-view/src/components/Attachments/Tabs/DatasetTabs/DatasetTabs.tsx
@@ -8,16 +8,9 @@ import {
 import { Locale } from '@epam/statgpt-shared-toolkit';
 import { IconButton } from '@epam/statgpt-ui-components';
 import classNames from 'classnames';
-import {
-  FC,
-  ReactNode,
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
+import { FC, ReactNode, useCallback, useEffect, useRef, useState } from 'react';
 import DatasetTab from './DatasetTab';
+import DatasetTabsCrossMode from './DatasetTabsCrossMode';
 import { useAdvancedView } from '../../../../context/AdvancedViewContext';
 import { Tooltip } from '../../../Tooltip/Tooltip';
 import { getTooltipDataByElement } from '../../../../utils/get-tooltip-data.by-element';
@@ -36,8 +29,6 @@ interface Props {
   selectDataset?: (datasetUrn?: string) => void;
   onOpenAdvancedView?: () => void;
 }
-
-const MAX_TABS_COUNT_CROSS_DATASET = 3;
 
 const DatasetTabs: FC<Props> = ({
   datasets,
@@ -112,48 +103,6 @@ const DatasetTabs: FC<Props> = ({
     [selectDataset],
   );
 
-  const datasetItems = useMemo(
-    () =>
-      (datasets || []).map((dataset) => ({
-        urn: generateShortUrn(dataset?.id, dataset?.version, dataset?.agencyID),
-        title: getLocalizedName(dataset, locale || Locale.EN),
-      })),
-    [datasets, locale],
-  );
-
-  const visibleDatasetItems = useMemo(() => {
-    if (datasetItems.length <= MAX_TABS_COUNT_CROSS_DATASET) {
-      return datasetItems;
-    }
-
-    const selectedDataset = datasetItems.find(
-      (dataset) => dataset.urn === initialSelectedDatasetUrn,
-    );
-
-    if (!selectedDataset) {
-      return datasetItems.slice(0, MAX_TABS_COUNT_CROSS_DATASET);
-    }
-
-    return [
-      selectedDataset,
-      ...datasetItems.filter((dataset) => dataset.urn !== selectedDataset.urn),
-    ].slice(0, MAX_TABS_COUNT_CROSS_DATASET);
-  }, [datasetItems, initialSelectedDatasetUrn]);
-
-  const hiddenDatasetsCount = Math.max(
-    datasetItems.length - visibleDatasetItems.length,
-    0,
-  );
-  const hiddenDatasetTitles = datasetItems
-    .filter(
-      (dataset) =>
-        !visibleDatasetItems.some(
-          (visibleDataset) => visibleDataset.urn === dataset.urn,
-        ),
-    )
-    .map((dataset) => dataset.title)
-    .join('\n');
-
   return (
     <div
       className={classNames(
@@ -165,25 +114,11 @@ const DatasetTabs: FC<Props> = ({
       )}
     >
       {isCrossDatasetModeOn ? (
-        <div className="dataset-tabs-list sm:w-[calc(100%-30px)]">
-          {visibleDatasetItems.map((dataset, index) => {
-            return (
-              <span key={dataset.urn} className="dataset-tabs-item-wrapper">
-                <span className="dataset-tabs-item" title={dataset.title}>
-                  {dataset.title}
-                </span>
-                {index < visibleDatasetItems.length - 1 && (
-                  <span className="dataset-tabs-separator" aria-hidden="true" />
-                )}
-              </span>
-            );
-          })}
-          {hiddenDatasetsCount > 0 && (
-            <span className="dataset-tabs-counter" title={hiddenDatasetTitles}>
-              +{hiddenDatasetsCount}
-            </span>
-          )}
-        </div>
+        <DatasetTabsCrossMode
+          datasets={datasets}
+          locale={locale}
+          initialSelectedDatasetUrn={initialSelectedDatasetUrn}
+        />
       ) : (
         <div className="flex items-center w-full overflow-y-auto gap-4 sm:w-[calc(100%-30px)]">
           {datasets?.map((dataset) => (
